pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "etsy-lister"
version = "1.0.0"
description = "Etsy Product Analysis & Recommendation System"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
requires-python = ">=3.9"
dependencies = [
    "beautifulsoup4>=4.12.0",
    "requests>=2.32.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "matplotlib>=3.9.0",
    "scikit-learn>=1.5.0",
    "wordcloud>=1.9.0",
    "flask>=3.0.0",
    "rich>=13.8.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "mypy>=1.11.0",
    "ruff>=0.6.0",
    "types-requests>=2.32.0",
]

[tool.ruff]
line-length = 140
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "C4", "SIM", "TCH"]
ignore = ["E501", "UP015", "B008", "B904"]

[tool.mypy]
python_version = "3.9"
warn_unused_ignores = true
warn_return_any = true
disallow_untyped_defs = true
ignore_missing_imports = true
warn_redundant_casts = true
warn_unused_configs = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-q --tb=short"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

